+++ src/components/molecule_action_group_card/index.buttons.tsx
@@ -1,0 +1,36 @@
+import { postActionByActionGroupId } from '@/api/action-groups/post-action-by-action-group-id.api'
+import StyledTextButtonAtom from '@/atoms/StyledTextButton'
+import { actionGroupFamily } from '@/recoil/action-groups/action-groups.state'
+import { Stack } from '@mui/material'
+import { FC } from 'react'
+import { useRecoilCallback, useRecoilValue } from 'recoil'
+
+interface Props {
+  id: string
+}
+const ActionGroupCardButton: FC<Props> = ({ id }) => {
+  const actionGroup = useRecoilValue(actionGroupFamily(id))
+
+  const onPost = useRecoilCallback(
+    // TODO: Write a hook with loading state
+    ({ set }) =>
+      async () => {
+        try {
+          const [data] = await postActionByActionGroupId(id)
+          set(actionGroupFamily(data.props.id), data)
+        } catch {}
+      },
+    [id],
+  )
+
+  if (!actionGroup?.isOpened || actionGroup.isTodayHandled) return null
+
+  return (
+    <Stack alignItems={`center`} direction={`row`}>
+      <StyledTextButtonAtom title="Post Action" onClick={onPost} />
+      <StyledTextButtonAtom title="Cannot commit today" />
+    </Stack>
+  )
+}
+
+export default ActionGroupCardButton
