+++ src/components/molecule_action_group_card/index.tsx
@@ -1,19 +1,15 @@
 import { FC, useCallback } from 'react'
-import { Card, Stack, Typography } from '@mui/material'
+import { Card, Stack } from '@mui/material'
 import StyledCloudRefresher from '@/atoms/StyledCloudRefresher'
 import { useActionGroupById } from '@/hooks/action-group/use-action-group-by-id.hook'
-import { useRecoilCallback, useRecoilValue } from 'recoil'
-import { actionGroupFamily } from '@/recoil/action-groups/action-groups.state'
 import ActivityCalendarById from '../molecule_activity_calendar/index.by-id'
-import StyledTextButtonAtom from '@/atoms/StyledTextButton'
-import { postActionByActionGroupId } from '@/api/action-groups/post-action-by-action-group-id.api'
-import { timeHandler } from '@/handlers/time.handler'
+import ActionGroupCardButton from './index.buttons'
+import ActionGroupCardTitle from './index.title'
 
 interface Props {
   id: string
 }
 const ActionGroupCard: FC<Props> = ({ id }) => {
-  const actionGroup = useRecoilValue(actionGroupFamily(id))
   const onGetActionGroupById = useActionGroupById(id)
 
   const onClickRefresh = useCallback(async () => {
@@ -21,42 +17,17 @@ const ActionGroupCard: FC<Props> = ({ id }) => {
     await Promise.all([onGetActionGroupById()])
   }, [onGetActionGroupById])
 
-  const onPost = useRecoilCallback(
-    // TODO: Write a hook with loading state
-    ({ set }) =>
-      async () => {
-        try {
-          const [data] = await postActionByActionGroupId(id)
-          set(actionGroupFamily(data.props.id), data)
-        } catch {}
-      },
-    [id],
-  )
-
   return (
     <Card>
       <Stack m={3}>
         {/* Header */}
         <Stack alignItems="center" direction={`row`} spacing={0.5} m={2}>
-          <Typography
-            variant="h6"
-            fontStyle={`italic`}
-            fontFamily={`Cormorant Garamond`}
-          >
-            {actionGroup
-              ? `"${actionGroup.props.task}" between ${timeHandler.getPrettyDate(actionGroup.props.openAt)} ~ ${timeHandler.getPrettyDate(actionGroup.props.closeAt)}`
-              : `Unknown Consistency`}
-          </Typography>
+          <ActionGroupCardTitle id={id} />
           <StyledCloudRefresher onClick={onClickRefresh} runOnClickOnce />
         </Stack>
         {/* Body */}
-        {!actionGroup?.isOpened && <ActivityCalendarById id={id} />}
-        {actionGroup?.isOpened && (
-          <Stack alignItems={`center`} direction={`row`}>
-            <StyledTextButtonAtom title="Post Action" onClick={onPost} />
-            <StyledTextButtonAtom title="Cannot commit today" />
-          </Stack>
-        )}
+        <ActivityCalendarById id={id} />
+        <ActionGroupCardButton id={id} />
         {/* Dialog */}
       </Stack>
     </Card>
