+++ src/components/molecule_activity_calendar/index.by-id.tsx
@@ -9,23 +9,21 @@ interface Props {
   id: string
 }
 const ActivityCalendarById: FC<Props> = ({ id }) => {
-  const actionGroupDailyPostWordChallenge = useRecoilValue(
-    actionGroupFamily(id),
-  )
+  const actionGroup = useRecoilValue(actionGroupFamily(id))
 
-  if (actionGroupDailyPostWordChallenge === undefined) return null
-  if (actionGroupDailyPostWordChallenge === null)
-    return <ActivityCalendarUnknown />
+  if (actionGroup === null) return <ActivityCalendarUnknown />
+  if (actionGroup === undefined) return null
+  if (actionGroup.isOpened && !actionGroup.isTodayHandled) return null
 
   return (
     <ReactActivityCalendar
-      loading={!actionGroupDailyPostWordChallenge}
+      loading={!actionGroup}
       theme={{
         light: [`#d4e6cf`, `#a7d4bf`, `#80d1ab`, `#56d197`, `#29cc7f`],
         dark: [`#d4e6cf`, `#a7d4bf`, `#80d1ab`, `#56d197`, `#29cc7f`],
       }}
-      totalCount={actionGroupDailyPostWordChallenge.totalCount}
-      data={actionGroupDailyPostWordChallenge.actions.map((p) => {
+      totalCount={actionGroup.totalCount}
+      data={actionGroup.actions.map((p) => {
         return {
           date: p.yyyymmdd,
           count: 1,
